22nd commit: The averages for systolic and diastolic are now correctly rounded to the nearest int. Used Math.round. Still need to fix a few minor bugs

+++ app/src/main/java/com/example/android/bpmonitor/model/BPStatus.java
@@ -7,7 +7,7 @@ public class BPStatus{
 
     private String mBPStatus = "";
 
-    public String getBPStatus(int s, int d){
+    public String getBPStatus(double s, double d){
         if((s>120)||(d>80)){
 
             mBPStatus = "high                ";
